requirements.txt
py-cord==2.6.1
PyNaCl==1.5.0
python-dotenv==1.0.1